More...

+++ www/admin/refereeplan.state.club.php
@@ -1,25 +1,27 @@
 <?php 
-case "clubteamsdbbf":
+case "clubteamsexternal":
+
+  $currentUser = mysql_fetch_assoc(getCurrentUser());
 
   echo '<script type="text/javascript" src="js/club.js"></script>';
   
   if($_POST["removeTeam"]){
   
-    ref_mysql_query("DELETE FROM `calendars` WHERE `id`='".$_POST["removeTeam"]."'");
+    ref_mysql_query("DELETE FROM `calendars` WHERE `id`='".$_POST["removeTeam"]."' AND `clubid`='".$currentUser['clubid']."'");
     $warning = fetchText("Team/League was removed.");
   
   }
   
   if($_POST["removeAllTeams"]){
     
-    ref_mysql_query("DELETE FROM `calendars`");
+    ref_mysql_query("DELETE FROM `calendars` WHERE `clubid`='".$currentUser['clubid']."'");
     $warning = fetchText("All of the Teams and Leagues has been removed.");
     
   }
   
   if($_POST["addAllTeams"]){
     
-    $newteams = addAllTeams();
+    $newteams = addAllTeams($currentUser['clubid']);
     
     if($newteams==0){
       $info = fetchText("No new Teams or Leagues.");
@@ -30,7 +32,7 @@ case "clubteamsdbbf":
   
   }
   
-  echo fetchText("Teams from DBBF","header2");
+  echo fetchText("External Teams","header2");
   echo showMessages($info,$warning,$error);
   echo '<br><a href="javascript:addAllTeams()">';
   echo fetchText("Add all of the Clubs Teams and Leagues");
@@ -39,7 +41,7 @@ case "clubteamsdbbf":
   echo fetchText("Remove all of the Clubs Teams and Leagues");
   echo '</a><br><br>';
   
-  $query = ref_mysql_query("SELECT * FROM `calendars` ORDER BY `team`");
+  $query = ref_mysql_query("SELECT * FROM `calendars` WHERE `clubid`='".$currentUser['clubid']."' ORDER BY `team`");
   
   echo fetchText("Teams:","header3");
   
@@ -62,10 +64,12 @@ break;
 
 case "clubteams":
 
+  require("class/refereeplan.class.club.php");
+
   echo '<script type="text/javascript" src="js/club.js"></script>';
   echo '<link rel="stylesheet" type="text/css" href="css/club.css">';
   
-  echo fetchText("Teams","header2");
+  echo fetchText("Club Teams","header2");
   echo '<div id="newTeamPlaceHolder" title="'.fetchText("New Team").'">
 	  <div id="teamNameHolder">'.fetchText("Team Name:").'</div><input id="newTeamName">
 	  <div id="teamContactHolder">'.fetchText("Contact:").'</div>
@@ -83,14 +87,26 @@ case "clubteams":
   $query = ref_mysql_query("SELECT * FROM `teams` ORDER BY `name` ASC");
 
   while($row = mysql_fetch_assoc($query)){
-    if($row['name']!="-"){
+  
+      $teams[] = new teamObj($row);
+    /*if($row['name']!="-"){
 	echo $row['name'];
 	echo ' - <a href="javascript:teamsChangeUser('.$row['id'].',\''.$row['name'].'\')">'.fetchText("Change Contact").'</a>';
 	echo ' - <a href="javascript:teamsChangeName('.$row['id'].',\''.$row['name'].'\')">'.fetchText("Change Name").'</a>';
 	echo ' - <a href="javascript:teamsRemove('.$row['id'].')">Fjern</a>';	
 	echo "<br>";
-    }
-}
+    }*/
+   }
+   
+   echo '<ul class="teamList">';
+		
+        foreach($teams as $team){
+	      echo $team;
+	}
+	
+
+   echo '</ul>';
+   
 break;
 
 ?>
